+++ packages/plugin/src/Extension/Crumb.php
@@ -42,13 +42,14 @@ final class Crumb extends CMSPlugin
         }
 
         $settings = [
-            'server'        => (string) $this->params->get('server', ''),
-            'service_body'  => (string) $this->params->get('service_body', ''),
-            'format_ids'    => (string) $this->params->get('format_ids', ''),
-            'view'          => (string) $this->params->get('view', ''),
-            'css_template'  => (string) $this->params->get('css_template', ''),
-            'base_path'     => (string) $this->params->get('base_path', ''),
-            'widget_config' => (string) $this->params->get('widget_config', ''),
+            'server'             => (string) $this->params->get('server', ''),
+            'service_body'       => (string) $this->params->get('service_body', ''),
+            'format_ids'         => (string) $this->params->get('format_ids', ''),
+            'view'               => (string) $this->params->get('view', ''),
+            'css_template'       => (string) $this->params->get('css_template', ''),
+            'base_path'          => (string) $this->params->get('base_path', ''),
+            'geolocation_radius' => (string) $this->params->get('geolocation_radius', ''),
+            'widget_config'      => (string) $this->params->get('widget_config', ''),
         ];
 
         $article->text = preg_replace_callback(